chore(set_llm): moved `set_llm` to new `utility_funcs` module
- Moved a utility function set_llm() to a new module. This function sets the selected model in the session state and displays an alert based on the model chosen. Creating a module for utility functions will help clean up the entry point file and make the entire project more readable, modular, and maintainable
