+++ freestream/main.py
@@ -14,6 +14,7 @@ from langchain_core.prompts import PromptTemplate
 from langchain.chains import ConversationalRetrievalChain
 from langchain_community.vectorstores import FAISS
 from langchain.text_splitter import RecursiveCharacterTextSplitter
+from utility_funcs import set_llm
 
 # Initialize LangSmith tracing
 os.environ["LANGCHAIN_TRACING_V2"] = "true"
@@ -148,17 +149,6 @@ model_names = {
         streaming=True
     ),
 }
-
-# Define a callback function for when a model is selected
-def set_llm():
-    # Set the model in session state
-    st.session_state.llm = model_names[selected_model]
-    
-    # Show an alert based on what model was selected
-    if st.session_state.model_selector == model_names["ChatOpenAI GPT-3.5 Turbo"]:
-        st.warning(body="Switched to ChatGPT 3.5-Turbo!", icon="⚠️")
-    else:
-        st.warning(body="Failed to change model! \nPlease contact the website builder.", icon="⚠️")
 
 selected_model = st.selectbox(
     label="Choose your chat model:",
